+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KS74DMJEY62AKJFN1GZ167HG"
+        "01KS7AZ80R6MVZHFH367H0FNWD"
     )
     r"""The unique identifier of the data source"""