+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KS74DMJK1THMQ0SYC616GHFJ"
+        "01KS7AZ80W4JJ49FHDDN0VTB1N"
     )
     r"""The unique identifier of the data source"""