+++ packages/orq-rc/RELEASES.md
@@ -8129,3 +8129,13 @@ Based on:
 - [python v4.10.0-rc.27] packages/orq-rc
 ### Releases
 - [PyPI v4.10.0-rc.27] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.27 - packages/orq-rc
+
+## 2026-05-22 07:59:25
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.764.0 (2.885.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.10.0-rc.28] packages/orq-rc
+### Releases
+- [PyPI v4.10.0-rc.28] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.28 - packages/orq-rc